feat: handle intial handshake b/w client and server
