+++ pkg/core/proxy/integrations/mysql/connection/authMoreDataPacket.go
@@ -2,11 +2,15 @@
 
 package connection
 
-import "go.keploy.io/server/v2/pkg/models/mysql"
+import (
+	"context"
+
+	"go.keploy.io/server/v2/pkg/models/mysql"
+)
 
 //ref: https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_connection_phase_packets_protocol_auth_more_data.html
 
-func DecodeAuthMoreData(data []byte) (*mysql.AuthMoreDataPacket, error) {
+func DecodeAuthMoreData(_ context.Context, data []byte) (*mysql.AuthMoreDataPacket, error) {
 	return &mysql.AuthMoreDataPacket{
 		StatusTag: data[0],
 		Data:      string(data[1:]),
